refactor: move lib to helpers
More descriptive of function now that services have moved away
