+++ src/screens/Collections/Item.tsx
@@ -2,7 +2,7 @@ import { Pressable, PressableProps, StyleSheet } from 'react-native'
 import Icon from 'react-native-vector-icons/MaterialCommunityIcons'
 
 import Typography from '../../components/Typography'
-import { OS } from '../../lib/consts'
+import OS from '../../helpers/os'
 import Colors from '../../themes/colors'
 import Fonts from '../../themes/fonts'
 import Units from '../../themes/units'
