+++ src/screens/Gurbani/Line.tsx
@@ -2,9 +2,9 @@ import { toUnicode } from 'gurmukhi-utils'
 import { StyleSheet, View } from 'react-native'
 
 import Typography from '../../components/Typography'
-import { OS } from '../../lib/consts'
-import Languages from '../../lib/languages'
-import transliterators from '../../lib/transliterators'
+import Languages from '../../helpers/languages'
+import OS from '../../helpers/os'
+import transliterators from '../../helpers/transliterators'
 import Colors from '../../themes/colors'
 import Fonts from '../../themes/fonts'
 import Units from '../../themes/units'
